% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_a_run.R
\name{analysis_a_run_server}
\alias{analysis_a_run_server}
\title{analysis_a_run_server}
\usage{
analysis_a_run_server(
  input,
  output,
  session,
  user,
  is_admin,
  signal,
  cache = TRUE
)
}
\description{
analysis_a_run_server
}
