% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxSidebarTest.R
\name{boxSidebarTest}
\alias{boxSidebarTest}
\title{boxSidebarTest}
\usage{
boxSidebarTest(
  ...,
  id = NULL,
  width = 50,
  background = "#333a40",
  startOpen = FALSE,
  icon = shiny::icon("cogs"),
  easyClose = TRUE
)
}
\description{
boxSidebarTest
}
