% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_padjust.R
\name{contrast_padjust}
\alias{contrast_padjust}
\title{contrast_padjust}
\usage{
contrast_padjust(
  model,
  contrast_list,
  data,
  variable,
  analysis_type = "Confirmatory"
)
}
\description{
contrast_padjust
}
