% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_output.R
\name{final_output}
\alias{final_output}
\title{final_output}
\usage{
final_output(
  transformed_data,
  toi,
  emmeans_obj,
  final_contrast,
  power,
  variable,
  save = "No",
  analysis_type
)
}
\description{
final_output
}
