% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate__for_Alex.R
\name{generate_data}
\alias{generate_data}
\title{generate_data}
\usage{
generate_data(
  group_specs,
  num_times,
  num_tech_reps,
  transform = 1,
  correlation = "ar",
  rho = 0.5,
  intercept = 5
)
}
\description{
generate_data
}
