% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_output.R
\name{html_table_gt}
\alias{html_table_gt}
\title{html_table_gt}
\usage{
html_table_gt(
  data,
  title,
  footer,
  include_summary,
  summary_only,
  transformation,
  analysis_type
)
}
\description{
html_table_gt
}
