% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_home.R
\name{make_test}
\alias{make_test}
\title{make_test}
\usage{
make_test(
  ns,
  test_name = "TEST 1",
  details = c("Single Treatment with Multiple Doses", "Multiple Time Points",
    "Controls and Comparators")
)
}
\description{
make_test
}
