% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_type_assignment_table.R
\name{make_type_assignment_table}
\alias{make_type_assignment_table}
\title{make_type_assignment_table}
\usage{
make_type_assignment_table(type_inputs, ns)
}
\description{
make_type_assignment_table
}
