% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_tables.R
\name{word_tables}
\alias{word_tables}
\title{word_tables}
\usage{
word_tables(
  data,
  include_summ_stat,
  transform,
  summary_only,
  footer_i = 1,
  footer_j = 1,
  footer = "footer",
  endpoint
)
}
\description{
word_tables
}
