% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_plot.R
\name{prism_plot}
\alias{prism_plot}
\title{prism_plot}
\usage{
prism_plot(
  data,
  tables,
  trt_sel = input$treatmentPlotSelectors,
  time_sel = input$timePlotSelectors,
  endpoint,
  format,
  cfb,
  power,
  box_width = 2,
  axis_title_size = 30,
  axis_text_size = 24,
  top_height = 1,
  bottom_height = 4,
  num_groups,
  type = "box",
  inputs = NULL
)
}
\description{
prism_plot
}
